feat: Refactor some function to utils file
