+++ client/src/utils/editorUtils.ts
@@ -1,3 +1,5 @@
+import * as monaco from 'monaco-editor';
+
 export function formatLanguage(language: string) {
   switch (language) {
     case 'javascript':
@@ -41,3 +43,54 @@ export function formatTime(ms: number): string {
     ? `${displayHoursString}:${displayMinutesString}:${displaySecondsString}`
     : `${displayMinutesString}:${displaySecondsString}`;
 }
+
+export function getLanguageId(language: Language): string | null {
+  const languageMapping: Record<Language, string> = {
+    javascript: '93',
+    python: '70',
+    java: '91',
+    csharp: '51',
+    cpp: '76',
+    ruby: '72',
+    go: '95',
+  };
+
+  return languageMapping[language];
+}
+
+export function calculateTotalPauseTime(
+  actionTimestamp: number,
+  endTimestamp: number,
+  recorderActions: RecorderActions
+) {
+  let totalPauseTime = 0;
+  for (let i = 0; i < recorderActions.pauseArray.length; i++) {
+    const pauseTimestamp = recorderActions.pauseArray[i].timestamp;
+    const resumeTimestamp = recorderActions.resumeArray[i]
+      ? recorderActions.resumeArray[i].timestamp
+      : endTimestamp;
+
+    if (actionTimestamp < pauseTimestamp) {
+      break;
+    }
+
+    if (actionTimestamp > resumeTimestamp) {
+      totalPauseTime += resumeTimestamp - pauseTimestamp;
+    } else {
+      totalPauseTime += actionTimestamp - pauseTimestamp;
+    }
+  }
+  return totalPauseTime;
+}
+
+// Set current theme based on if darkmode is on or not
+export const toggleTheme = (
+  darkMode: string,
+  monacoInstance: typeof monaco
+) => {
+  if (darkMode) {
+    monacoInstance.editor.setTheme('vs-dark');
+  } else {
+    monacoInstance.editor.setTheme('vs-light');
+  }
+};
